requirement.txt
pandas==2.2.2
requests==2.32.3
psycopg2-binary==2.9.9
PyYAML==6.0.2
apache-airflow==2.9.3
